add a null check, just in case.
Closes #466

+++ src/main/java/vazkii/psi/common/core/handler/PersistencyHandler.java
@@ -88,7 +88,7 @@ public final class PersistencyHandler {
 			return;
 
 		try {
-			if (persistentFile.exists() || persistentFile.createNewFile()) {
+			if (persistentFile != null && (persistentFile.exists() || persistentFile.createNewFile())) {
 				try (BufferedWriter writer = new BufferedWriter(new FileWriter(persistentFile))) {
 					writer.write("" + level);
 					writer.flush();
